+++ examples/provider/provider.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     quicknode = {
       source  = "quicknode/quicknode"
-      version = "~> 0.1"
+      version = "~> 0.3.0"
     }
   }
 }